Fixing The Jump
I changed to code of the jump so that when the character is less that
frame 42 or equal to 78 the character will jump. I also moved the walk
cycle handler into the character boundary handler because it was in an
enter frame which was continuously being programmed which stopped the
character from jumping.